update avatar test scene
