+++ examples/template.tsx
@@ -5,7 +5,7 @@ import { LoadEngineWithScene } from '@xrengine/client-core/src/components/World/
 import OfflineLocation from '@xrengine/client-core/src/components/World/OfflineLocation'
 import { LocationAction } from '@xrengine/client-core/src/social/services/LocationService'
 import { useDispatch } from '@xrengine/client-core/src/store'
-import { useEngineState } from '@xrengine/engine/src/ecs/classes/EngineService'
+import { useEngineState } from '@xrengine/engine/src/ecs/classes/EngineState'
 
 import Layout from '@xrengine/client-core/src/components/Layout'
 import { loadSceneJsonOffline } from '@xrengine/client/src/pages/offline/utils'
